Edits to get the library to compile on Windows:
Updated the CMakeLists file to use find_package() to find MKL, SuiteSparse, and OpenMP.
Put in some workarounds for platform-specific headers and macros.
Changed "and" to "&&".
Changed schedule(auto) to schedule(static) for OpenMP loops. This should be equivalent on GCC; auto is not supported on Windows.
Changed some functions with return type int, that do not actually return a value, to return type void.